requirements.txt
Django
django-allauth
gunicorn
mysqlclient
Pillow
python-decouple
stripe
uuid